Added zoom functionality to the 'X' and 'Z' keys and adjusted pan speed
Zoom functionality now added for debug mode. KeyboardCameraInputCompo-
-nent now handles 'X', calling the zoomOut function in CameraActions,
which adds to the OrthographicCamera's zoom variable until it reaches
the value of 2. KeyboardCameraInputComponent also handles 'Z'. It calls
the vfunctio zoomIn which subtracts from the camera's zoom value, until
it reaches 0.

Panning has been slowed to a third of the previous speed.